propTypes validated
