+++ src/components/InputTodo.jsx
@@ -1,0 +1,42 @@
+import { useState } from 'react';
+import PropTypes from 'prop-types';
+
+const InputTodo = ({ addTodoItem }) => {
+  const [title, setTitle] = useState('');
+  const [message, setMessage] = useState('');
+
+  const handleChange = (e) => {
+    setTitle(e.target.value);
+  };
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (title.trim()) {
+      addTodoItem(title);
+      setTitle('');
+      addTodoItem(title);
+      setMessage('');
+    } else {
+      setMessage('Please add item.');
+    }
+  };
+  return (
+    <>
+      <form onSubmit={handleSubmit}>
+        <input
+          type="text"
+          placeholder="Add Todo..."
+          value={title}
+          onChange={handleChange}
+        />
+        <button type="button">Submit</button>
+      </form>
+      <span>{message}</span>
+    </>
+  );
+};
+
+InputTodo.propTypes = {
+  addTodoItem: PropTypes.func.isRequired,
+};
+
+export default InputTodo;
