+++ src/components/TodoItem.jsx
@@ -1,0 +1,39 @@
+import PropTypes from 'prop-types';
+
+const TodoItem = ({ itemProp, setTodos, delTodo }) => {
+  const handleChange = (id) => {
+    setTodos((prevState) => prevState.map((todo) => {
+      if (todo.id === id) {
+        return {
+          ...todo,
+          completed: !todo.completed,
+        };
+      }
+      return todo;
+    }));
+  };
+
+  return (
+    <li>
+      <input
+        type="checkbox"
+        checked={itemProp.completed}
+        onChange={() => handleChange(itemProp.id)}
+      />
+      <button type="button" onClick={() => delTodo(itemProp.id)}>Delete</button>
+      {itemProp.title}
+    </li>
+  );
+};
+
+TodoItem.propTypes = {
+  itemProp: PropTypes.shape({
+    id: PropTypes.number.isRequired,
+    completed: PropTypes.bool.isRequired,
+    title: PropTypes.string.isRequired,
+  }).isRequired,
+  setTodos: PropTypes.func.isRequired,
+  delTodo: PropTypes.func.isRequired,
+};
+
+export default TodoItem;
