+++ src/components/TodosList.jsx
@@ -1,0 +1,28 @@
+import PropTypes from 'prop-types';
+import TodoItem from './TodoItem';
+
+const TodosList = ({ todosProps, setTodos, delTodo }) => (
+  <ul>
+    {todosProps.map((todo) => (
+      <TodoItem
+        key={todo.id}
+        itemProp={todo}
+        setTodos={setTodos}
+        delTodo={delTodo}
+      />
+    ))}
+  </ul>
+);
+
+TodosList.propTypes = {
+  todosProps: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      completed: PropTypes.bool.isRequired,
+    }),
+  ).isRequired,
+  setTodos: PropTypes.func.isRequired,
+  delTodo: PropTypes.func.isRequired,
+};
+
+export default TodosList;
