+++ src/index.js
@@ -4,15 +4,13 @@ import reactLogo from './images/react.svg';
 import './styles/app.scss';
 import TodosLogic from './components/TodosLogic';
 
-
 const rootElement = document.getElementById('root');
 
 const root = ReactDOM.createRoot(rootElement);
 root.render(
-    <div>
-        <TodosLogic />
-        <h1>Hello from React application</h1>
-        <img src={reactLogo} className="logo react" alt="React logo" />
-    </div>
+  <div>
+    <TodosLogic />
+    <h1>Hello from React application</h1>
+    <img src={reactLogo} className="logo react" alt="React logo" />
+  </div>,
 );
-
