+++ src/styles/app.scss
@@ -1,4 +1,3 @@
 body {
-    color: #ff2778;
-  }
-  
+  color: #ff2778;
+}
